backend/AI/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
appdirs==1.4.4
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
datasets==2.21.0
dill==0.3.8
esprima==4.0.1
fasttext==0.9.3
filelock==3.15.4
Flask==3.0.3
Flask-Cors==4.0.1
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.6.1
huggingface-hub==0.24.6
humanfriendly==10.0
idna==3.7
importlib_metadata==8.2.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.2.1
numpy==1.26.4
onnx==1.16.2
onnxruntime==1.19.0
optimum==1.21.4
packaging==24.1
pandas==2.2.2
protobuf==5.27.3
pyarrow==17.0.0
pybind11==2.13.5
pyee==11.1.0
pyppeteer==2.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
safetensors==0.4.4
sentencepiece==0.2.0
six==1.16.0
soupsieve==2.6
sympy==1.13.2
tokenizers==0.19.1
torch==2.4.0
tqdm==4.66.5
transformers==4.43.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.19
websockets==10.4
Werkzeug==3.0.3
xxhash==3.5.0
yarl==1.9.4
zipp==3.19.2
